includes distance time keyword into parsing
